requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
Django==1.7.5
django-filter==0.9.2
djangorestframework==3.0.5
docutils==0.12
gnureadline==6.3.3
ipython==3.0.0
Jinja2==2.7.3
jsonschema==2.4.0
Markdown==2.6
MarkupSafe==0.23
mistune==0.5
nose==1.3.4
numpy==1.9.2
numpydoc==0.5
pandas==0.15.2
psycopg2==2.6
ptyprocess==0.4
Pygments==2.0.2
python-dateutil==2.4.0
pytz==2014.10
pyzmq==14.5.0
requests==2.5.3
six==1.9.0
Sphinx==1.2.3
terminado==0.5
tornado==4.1
